Update realmofreckoning.py
I moved some of the methods in Location and Character classes into the Map class, also changing some of the Map methods into class methods.  Everything is once again functional as before and I believe the new organization of methods makes more sense than before, when we had to reference an instance of a Location to check if it existed (duh.)  I'll keep going through this program looking for ways to improve the logic and structure as well as add features.