move rewrite (now named keydown)
Character class' 'move' method was renamed 'keydown' and significantly reduced in length due to implementing a dictionary with keys made up of possible user input and values made up of the names of corresponding methods to call for each key pressed.  This was made possible by learning that you can append parentheses ( () ) to the end of a string which matches the name of a method and doing so will actually call that method.  Nice!